Move non-main files to inc/ directory; move all running code to bottom of CTRLDeck_Python.py script
